delete aof in pikiwi
