+++ clear.sh
@@ -10,6 +10,9 @@
 # ******************************************************
 
 rm -rf ./cmake-build-debug
+rm -rf ./deps
+rm -rf ./buildtrees
 rm -rf ./build
 rm -rf ./bin
+rm -rf ./dbsync
